float-steps: fix potential segmentation fault

+++ src/cpp/util/other/float-steps.cpp
@@ -8,7 +8,7 @@ int main()
 {
     for (const auto&& i : std::ranges::iota_view(0ull, ~0ull)) {
         float f;
-        reinterpret_cast<std::uint64_t&>(f) = i;
+        reinterpret_cast<std::uint32_t&>(f) = i;
         std::cout << f << '\n';
     }
